+++ main4.py
@@ -2,9 +2,11 @@ import pickle
 import os
 import numpy
 from collections import Counter
+import collections
 from model.entity_quotes import *
 from model.complete_quote import *
 from model.utils import *
+from pathlib import Path
 from itertools import * 
 from sentence_transformers import SentenceTransformer
 from scipy.spatial import distance
@@ -36,11 +38,41 @@ def getEmbs(completeEntity):
     size = 1024
     #quote_batch, n_of_quotes = get_all_entity_quotes(completeEntity)
     all = []
+    texts = []
+    remember = []
     for language in completeEntity.entities:
         quotes = list(completeEntity.entities[language][0].quotes.values())
         for quote in quotes:
             all.append(quote) 
-    texts = [i.quote for i in all]
+    for j, i in enumerate(all):
+        if hasattr(i, "quote"):
+            if hasattr(i.quote, "text"):
+                texts.append(i.quote.text)
+            else:
+                texts.append(i.quote)
+        elif hasattr(i, "original"):
+            if hasattr(i.original, "text"):
+                texts.append(i.original.text)
+            else:
+                texts.append(i.original)
+        elif hasattr(i, "translation"):
+            if hasattr(i.translation, "text"):
+                texts.append(i.translation.text)
+            else:
+                texts.append(i.translation)
+        else:
+            # if we somehow got a quote object without quote, orginal or translation attribute. We want to delete it from the list
+            remember.append(j)
+            inv_map = {v: k for k, v in completeEntity.entitiesitems()}
+            for language in completeEntity.entities:
+                if j > len(list(completeEntity.entities[language][0].quotes.values())):
+                    j = j - len(list(completeEntity.entities[language][0].quotes.values()))
+                else:
+                    inv_map = {v: k for k, v in completeEntity.entities[language][0].quotes.items()}
+                    key = inv_map[i]
+                    del completeEntity.entities[language][0].quotes[key]           
+    for index in sorted(remember, reverse=True):
+        del all[index]
     #quotes.update(quote_batch)
     #if list(quotes.values()) == []:
     if all == []:
@@ -137,14 +169,11 @@ def X(intermediate_done=False):
     d ={}
     dir = "/home/kuculo/quotekg/v1_final"
     subdirs = [x[0] for x in os.walk(dir)][1:] 
-    main1 = ["/home/kuculo/quotekg/v1_final/it", "/home/kuculo/quotekg/v1_final/en"]
-    main2 = ["/home/kuculo/quotekg/v1_final/pl", "/home/kuculo/quotekg/v1_final/cs", "/home/kuculo/quotekg/v1_final/pt", "/home/kuculo/quotekg/v1_final/uk","/home/kuculo/quotekg/v1_final/fr"]
-    main3 = ["/home/kuculo/quotekg/v1_final/ru","/home/kuculo/quotekg/v1_final/es","/home/kuculo/quotekg/v1_final/fa","/home/kuculo/quotekg/v1_final/bg","/home/kuculo/quotekg/v1_final/de","/home/kuculo/quotekg/v1_final/he","/home/kuculo/quotekg/v1_final/zh","/home/kuculo/quotekg/v1_final/bs","/home/kuculo/quotekg/v1_final/ar"]
-    subdirs = list(set(subdirs) - set(main1)-set(main2)-set(main3))
+    subdirs = ["/home/kuculo/quotekg/v1_final/it", "/home/kuculo/quotekg/v1_final/en"]
     print("v2:")  
     if not os.path.isdir("/home/kuculo/quotekg/v2_final"):
-        os.mkdir("/home/kuculo/quotekg/v2_final")    
-    if not intermediate_done:                                                                      
+        os.mkdir("/home/kuculo/quotekg/v2_final")                                                                          
+    if not intermediate_done:
         for i, subdir in enumerate(subdirs):
             print(subdir)
             print("%d of %d complete"%(i, len(subdirs)))  
@@ -190,14 +219,20 @@ def X(intermediate_done=False):
                         if language not in d[filename]:
                             d[filename].update({language:[]})
                             d[filename][language].append(entity)
-        
     print("Creating CompleteEntities and embedding quotes")
-
-    for i, filename in enumerate(d):
-        print("%d of %d complete"%(i, len(d))) 
+    # od = collections.OrderedDict(sorted(d.items())[15000:20000])
+    od = collections.OrderedDict(sorted(d.items())[40000:45000])
+    # od = collections.OrderedDict(sorted(d.items())[65000:70000])
+    d = None
+    for i, filename in enumerate(od):
+        print("%d of %d complete"%(i, len(od))) 
+        path = "/home/kuculo/quotekg/v2_final/"+filename
+        path = Path(path)
+        if path.exists():
+            continue
         print(filename)
         with open("/home/kuculo/quotekg/v2_final/"+filename,"wb") as f:
-            new = CompleteEntity(filename[:-4], d[filename])
+            new = CompleteEntity(filename[:-4], od[filename])
             new = getEmbs(new)
             pickle.dump(new, f)
       
@@ -291,11 +326,14 @@ def smallTest2():
             print(quote.embedding)
 
 if __name__ == "__main__":
-    X()
+    X(intermediate_done=True)
+    while(True):
+        print("embedding completed")
     entity_dir = "/home/kuculo/quotekg/v2_final/"
     #subdirs = [x[0] for x in os.walk(entity_dir)][1:]  
     sim = 0.8
-    
+    print("Embeddings created")
+    print("Clustering")
     for root,dirs,files in os.walk(entity_dir):
         for z, filename in enumerate(files):
             with open(entity_dir+"/"+filename,"rb") as f:
